+++ src/cli/replay-test/reporters/default.ts
@@ -6,7 +6,7 @@ import {
 } from '../progress.ts';
 import { formatDurationSeconds } from '../duration-format.ts';
 import { colorize, supportsColor } from '../../../commands/output/color.ts';
-import { formatReplayDivergenceReport } from '@agent-device/contracts/divergence';
+import { formatReplayDivergenceReport } from '../../../core/replay-divergence.ts';
 import type {
   ReplayTestReporter,
   ReplayTestReporterContext,
