+++ src/daemon/client/daemon-client-lifecycle.ts
@@ -3,7 +3,7 @@ import net from 'node:net';
 import os from 'node:os';
 import path from 'node:path';
 import { AppError, normalizeError } from '@agent-device/kernel/errors';
-import { readReplayDivergenceResume } from '@agent-device/contracts/divergence';
+import { readReplayDivergenceResume } from '../../core/replay-divergence.ts';
 import type { DaemonRequest, DaemonResponse } from '../daemon-request.ts';
 import {
   runCmdDetachedMonitored,
