+++ src/daemon/replay/internal/session-replay-divergence-publication.ts
@@ -2,7 +2,8 @@ import fs from 'node:fs';
 import path from 'node:path';
 import type { ResponseLevel } from '@agent-device/kernel/contracts';
 import { redactDiagnosticData } from '@agent-device/kernel/redaction';
-import { boundReplayDivergence, type ReplayDivergence } from '@agent-device/contracts/divergence';
+import { type ReplayDivergence } from '@agent-device/contracts/divergence';
+import { boundReplayDivergence } from '../../../core/replay-divergence.ts';
 import {
   bindInternalObservationAuthority,
   type InternalObservationEvidence,
